Add user count and last events to dashboard
Updated UserServices to include a method for getting the number of users. Modified DashboardController to fetch and pass the number of users and the list of last events to the dashboard view. Updated ExceptionController to log exceptions and pass error messages to the error view. Added new images to the static resources. Updated dashboard.html to display the number of created users and the list of last created events. Fixed a minor whitespace issue in events.html.